Added runtime classes for crun/wasm/nvidia

Added default runtime flag
