+++ pkg/cli/cmds/agent.go
@@ -27,6 +27,7 @@ type Agent struct {
 	Snapshotter              string
 	Docker                   bool
 	ContainerRuntimeEndpoint string
+	DefaultRuntime           string
 	ImageServiceEndpoint     string
 	FlannelIface             string
 	FlannelConf              string
@@ -114,6 +115,11 @@ var (
 		Name:        "container-runtime-endpoint",
 		Usage:       "(agent/runtime) Disable embedded containerd and use the CRI socket at the given path; when used with --docker this sets the docker socket path",
 		Destination: &AgentConfig.ContainerRuntimeEndpoint,
+	}
+	DefaultRuntimeFlag = &cli.StringFlag{
+		Name:        "default-runtime",
+		Usage:       "(agent/runtime) Set the default runtime in containerd",
+		Destination: &AgentConfig.DefaultRuntime,
 	}
 	ImageServiceEndpointFlag = &cli.StringFlag{
 		Name:        "image-service-endpoint",
@@ -258,6 +264,7 @@ func NewAgentCommand(action func(ctx *cli.Context) error) cli.Command {
 			LBServerPortFlag,
 			ProtectKernelDefaultsFlag,
 			CRIEndpointFlag,
+			DefaultRuntimeFlag,
 			ImageServiceEndpointFlag,
 			PauseImageFlag,
 			SnapshotterFlag,
