+++ pkg/daemons/config/types.go
@@ -55,6 +55,7 @@ type Node struct {
 	Token                    string
 	Certificate              *tls.Certificate
 	ServerHTTPSPort          int
+	DefaultRuntime           string
 }
 
 type Containerd struct {
